% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integer-addition.R
\name{integer_addition}
\alias{integer_addition}
\title{adds two integers}
\usage{
integer_addition(a, b)
}
\arguments{
\item{a}{first integer}

\item{b}{second integer}
}
\description{
Adding two numbers
}
