% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integer-multiplication.R
\name{integer_multiplication}
\alias{integer_multiplication}
\title{multiply two integers}
\usage{
integer_multiplication(a, b)
}
\arguments{
\item{a}{first integer}

\item{b}{second integer}
}
\description{
multiply two integers
}
