% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integer-addition.R
\name{integer_addition}
\alias{integer_addition}
\title{adds three, three integers, AH! AH! AH!}
\usage{
integer_addition(a, b, c)
}
\arguments{
\item{a}{first integer}

\item{b}{second integer}

\item{c}{third integer}
}
\description{
Adding three numbers
}
